refactor(structure): index for simple imports
