+++ src/modules/Button/index.js
@@ -1,0 +1,3 @@
+import Button from './Button.component';
+
+export default Button;
