+++ src/modules/Card/index.js
@@ -1,0 +1,3 @@
+import Card from './Card.component';
+
+export default Card;
