+++ src/modules/Header/index.js
@@ -1,0 +1,3 @@
+import Header from './Header.component';
+
+export default Header;
