+++ src/pages/Home.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import Layout from '../modules/Layout/Layout.component';
-import Header from '../modules/Header/Header.component';
+import Header from '../modules/Header';
 import PokeDex from '../modules/PokeDex';
 
 const Home = () => (
